requirements.txt
streamlit
pandas
openpyxl
jupyter
requests
loguru
plotly
sentence-transformers
pytest
python-dotenv
sqlalchemy
alembic
psycopg2-binary
streamlit-authenticator
pyyaml

